+++ app/src/main/java/com/millicast/android_app/SubscribeFragment.java
@@ -27,7 +27,7 @@ import org.webrtc.RendererCommon;
 
 import static android.media.AudioManager.MODE_IN_COMMUNICATION;
 import static android.media.AudioManager.MODE_NORMAL;
-import static com.millicast.android_app.MillicastManager.Source.CURRENT;
+import static com.millicast.android_app.MCTypes.Source.CURRENT;
 import static com.millicast.android_app.Utils.logD;
 import static com.millicast.android_app.Utils.makeSnackbar;
 import static com.millicast.android_app.Utils.populateSpinner;